requirements.txt
absl-py==1.2.0
albumentations==1.3.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
astunparse==1.6.3
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.2.0
catboost==1.1.1
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
cycler==0.11.0
dbus-python==1.2.16
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
efficientnet-pytorch==0.7.1
entrypoints==0.4
executing==1.0.0
fastjsonschema==2.16.1
flatbuffers==2.0.7
fonttools==4.37.1
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
group-lasso==1.5.0
grpcio==1.48.1
h5py==3.7.0
idna==2.8
imageio==2.22.2
imbalanced-learn==0.9.1
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipykernel==5.1.1
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.17.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.15.0
jupyter==1.0.0
jupyter-client==7.3.5
jupyter-console==6.4.4
jupyter-contrib-core==0.4.0
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.11.1
jupyter-highlight-selected-word==0.2.0
jupyter-http-over-ws==0.0.8
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.5.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kaleido==0.2.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
lightgbm==3.3.3
llvmlite==0.39.1
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.4
munch==2.5.0
nbclient==0.6.7
nbconvert==7.0.0
nbformat==4.4.0
nest-asyncio==1.5.5
networkx==2.8.7
notebook==6.4.12
numba==0.56.4
numpy==1.23.2
oauthlib==3.2.0
opencv-python-headless==4.6.0.66
openslide-python==1.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil-resolve-name==1.3.10
plotly==5.11.0
pretrainedmodels==0.7.4
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.19.4
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
PyGObject==3.36.0
pynndescent==0.5.8
pyparsing==3.0.9
pyrsistent==0.18.1
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
pytz==2022.5
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==23.2.1
qtconsole==5.3.2
QtPy==2.2.0
qudida==0.0.4
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
segmentation-models-pytorch==0.3.0
Send2Trash==1.8.0
six==1.14.0
soupsieve==2.3.2.post1
stack-data==0.5.0
tenacity==8.1.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.15.0
threadpoolctl==3.1.0
tifffile==2022.10.10
timm==0.4.12
tinycss2==1.1.1
torch==1.13.0+cu116
torchaudio==0.13.0+cu116
torchsampler==0.1.2
torchsummary==1.5.1
torchvision==0.14.0+cu116
tornado==6.2
tqdm==4.64.1
traitlets==5.3.0
ttach==0.0.3
typing-extensions==4.3.0
umap-learn==0.5.3
urllib3==1.25.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
xgboost==1.7.1
zipp==3.8.1
